[#218] :sparkles: refactor: 문의 모달 폴더 구조화 및 그에 따른 경로 수정과 기존 파일 제거
