+++ client/src/components/Shared/Modal/RequestModal/RequestModal.tsx
@@ -2,8 +2,8 @@ import React, { useCallback, useState } from 'react';
 import Button from '@/components/Shared/Button';
 import { Input } from '@/components/Shared/Input';
 import Title from '@/components/Shared/Title';
-import ModalLayout from './ModalLayout';
-import * as S from './styles';
+import ModalLayout from '../ModalLayout';
+import * as S from '../styles';
 import { Textarea } from '@/components/Shared/Input';
 import Form from '@/components/Shared/Form';
 import CategorySelector from '@/components/CategorySelector';
